站点首页数据查询完成

+++ src/main/java/com/kaige/entity/Comment.java
@@ -19,7 +19,7 @@ public interface Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY
     )
-    BigInteger id();
+    int id();
 
     /**
      * 昵称
